example/requirements.txt
Flask-Script==0.6.7
